+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 [Unreleased]: https://github.com/rynkowsg/shell-gr/compare/v0.3.0..main
 
+- **install**: Fix installation on Apple M1
 ## [0.3.0](https://github.com/rynkowsg/shell-gr/commits/v0.3.0) (2024-04-18)
 
 - **normalize**: Add function to unset overridden `cd`